# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CallWeave API
  version: 0.1.0
  description: |
    CallWeave 自身の API。Vonage / ASR ベンダーのエンドポイントではない。
    /api/* は callweave-web（LB + Cloud Armor 配下）、/webhooks/* と /internal/* は callweave-control、
    /media/vonage は callweave-media が提供する。
    エラーは { code, message, requestId }。個人情報を含む応答は Cache-Control: no-store。
servers:
  - url: https://callweave.example.com
paths:
...
  /media/vonage:
    get:
      summary: Vonage 監視 WS のアップグレード
      description: Authorization に callId / role / 世代付きの有効期限トークン。古い世代・不正トークンは拒否。
      security: [{ mediaToken: [] }]
      responses: { '101': { description: Switching Protocols }, '401': { description: invalid token } }
...
components:
  securitySchemes:
    vonageSignature: { type: http, scheme: bearer, bearerFormat: JWT, description: Vonage 署名シークレットで HMAC 署名された JWT }
    mediaToken: { type: http, scheme: bearer, bearerFormat: JWT }
    googleOidc: { type: http, scheme: bearer, bearerFormat: JWT, description: Cloud Tasks / Scheduler の SA の OIDC トークン }
...